// src/types/mongoose-types.ts
export type TaskQueries = {};

export type TaskMethods = {};

export type TaskStatics = {};

/**
 * Mongoose Model type
 *
 * Pass this type to the Mongoose Model constructor:
 * ```
 * const Task = mongoose.model<TaskDocument, TaskModel>("Task", TaskSchema);
 * ```
 */
export type TaskModel = mongoose.Model<TaskDocument, TaskQueries> & TaskStatics;
// ...
export type TaskDocument = mongoose.Document<
  mongoose.Types.ObjectId,
  TaskQueries
> &
  TaskMethods & {
    title: string;
    description: string;
    status?: "pending" | "in-progress" | "completed";
    priority: "low" | "medium" | "high";
    dueDate: Date;
    userId: UserDocument["_id"] | UserDocument;
    _id: mongoose.Types.ObjectId;
    createdAt?: Date;
    updatedAt?: Date;
  };
// ...
export type UserQueries = {};

export type UserMethods = {};
// ...
export type UserDocument = mongoose.Document<
  mongoose.Types.ObjectId,
  UserQueries
> &
  UserMethods & {
    username: string;
    password: string;
    _id: mongoose.Types.ObjectId;
    createdAt?: Date;
    updatedAt?: Date;
  };